+++ CHANGES.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ## v3.x.x
+- VM: set the matplotlib backend to tkagg
 - Updated Versions:
   - CIL: 4f6e3cfb0648d08071974eba35c1369a892d4924
 - CMake: